* eval.c, file.c, gc.c, io.c, object.c, ruby.c, ruby.h, struct.c,   ext/socket/socket.c: differentiate long and int; use proper   printf type specifiers and do casts where appropriate.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@3014 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ ruby.c
@@ -930,7 +930,7 @@ set_arg0(val, id)
     }
     rb_progname = rb_tainted_str_new(s, i);
 #elif defined(HAVE_SETPROCTITLE)
-    setproctitle("%.*s", i, s);
+    setproctitle("%.*s", (int)i, s);
     rb_progname = rb_tainted_str_new(s, i);
 #else
     if (len == 0) {
